Player Position and Weapon Activation.
I Handle the Player Position by taking the mouse position and convert it from Screen Point Position to World Space Position. I added a clamp so the Spaceship won't go beyond the boundaries of the game window.
I wrote a simple method for checking if the two mouse buttons are being held independently.